// apps/ulixee/noderdom/src/Types.ts
export interface Typed {
	type: string | Typed[];
	subtype?: Typed | Typed[];
	nullable?: 1;
}

export interface Param extends Typed {
	name: string;
	optional?: 1;
	variadic?: 1;
}